fix issues w the parser
