+++ best/text/str.h
@@ -848,13 +848,14 @@ class pretext<E>::rune_index_iter_impl final {
   friend best::iter<rune_index_iter_impl>;
   friend best::iter<rune_index_iter_impl&>;
 
-  constexpr explicit rune_index_iter_impl(rune_try_iter iter) : iter_(iter) {}
+  constexpr explicit rune_index_iter_impl(rune_try_iter iter)
+      : iter_(iter), size_(iter->rest().size()) {}
   constexpr best::option<best::row<size_t, best::rune>> next();
   constexpr best::size_hint size_hint() const { return iter_.size_hint(); }
   constexpr size_t count() && { return BEST_MOVE(iter_).count(); }
 
   rune_try_iter iter_;
-  size_t idx_ = 0;
+  size_t size_;
 };
 
 template <typename E>
@@ -1157,13 +1158,10 @@ text<E>::rune_index_iter_impl::next() {
 template <typename E>
 constexpr best::option<best::row<size_t, best::rune>>
 pretext<E>::rune_index_iter_impl::next() {
-  size_t cur_len = iter_->rest().size();
   auto next = iter_.next();
   BEST_GUARD(next);
 
-  size_t idx = idx_;
-  idx_ += cur_len - iter_->rest().size();
-  return {{idx, next->ok().value_or(rune::Replacement)}};
+  return {{size_ - rest().size(), next->ok().value_or(rune::Replacement)}};
 }
 template <typename E>
 constexpr best::option<best::result<best::rune, best::encoding_error>>
